requirementx.txt
amqp==5.0.9
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.11
django-celery-beat==2.2.1
django-timezone-field==4.2.3
django-widget-tweaks==1.4.12
kombu==5.2.3
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.24
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
